Simplified todolist example
